frontend/server/requirements.txt
playwright==1.39.0
beautifulsoup4==4.12.2
Pillow==10.1.0
numpy==1.26.2
insightface==0.7.3
torch==2.1.1
transformers==4.35.2
python-dotenv==1.0.0
flask==2.3.3
flask-cors==4.0.0
flask-socketio==5.3.6
python-dotenv==1.0.0
